transcriber-app/requirements.txt
pydub
resemblyzer
scikit-learn
numpy
psutils
